modified orders of user page, create function get order list by user id use like parameter search, modified order hook, order service, change prettier config file in front end, bring dto from backend to front end to synchronize dto, create 2 new component are order item and order list

+++ backend/src/core/order/entites/orders.entity.ts
@@ -27,7 +27,7 @@ export class Order {
   @AutoMap()
   user: User;
 
-  @Column('int', { name: 'total_price' })
+  @Column('double', { name: 'total_price' })
   @AutoMap()
   totalPrice: number;
